+++ frontend/src/screens/HomeScreen.js
@@ -1,9 +1,9 @@
 import React from "react";
-import HomePage from "../shared/components/homePage";
 // import "./HomeScreen.css"; // Import the CSS file for styling
 import "../styles/HomeScreen.css";
 import { useSelector } from "react-redux";
-import DashBoard from '../shared/components/dashboard'
+import HomePage from '../shared/components/homePage'
+import DashBoard from "../shared/components/dashBoard";
 const HomeScreen = () => {
   const { userInfo, loading } = useSelector((state) => state.auth);
 
@@ -13,7 +13,7 @@ const HomeScreen = () => {
 
   return (
     <div className="home-screen">
-      {userInfo ? <HomePage /> : <DashBoard />}
+      {userInfo ? <DashBoard /> : <HomePage/>}
     </div>
   );
 };
